some refactoring + nx migration
