+++ apps/frontend/project.json
@@ -31,12 +31,6 @@
               "maximumError": "4kb"
             }
           ],
-          "fileReplacements": [
-            {
-              "replace": "libs/shared/environments/src/lib/environment.ts",
-              "with": "libs/shared/environments/src/lib/environment.prod.ts"
-            }
-          ],
           "outputHashing": "all"
         },
         "development": {
