+++ libs/shared/data-access/database/src/index.ts
@@ -1,2 +1,2 @@
 export * from './lib/shared-data-access-database.module';
-export { SupabaseService } from './lib/supabase.service';
+export * from './lib/supabase.service';
