+++ package.json
@@ -25,6 +25,7 @@
     "@nestjs/core": "^9.0.0",
     "@nestjs/platform-express": "^9.0.0",
     "@ngneat/tailwind": "^7.0.3",
+    "@nrwl/devkit": "14.5.10",
     "@supabase/supabase-js": "^1.35.6",
     "daisyui": "^2.24.0",
     "reflect-metadata": "^0.1.13",
@@ -42,17 +43,17 @@
     "@angular/language-service": "~14.1.1",
     "@nestjs/schematics": "^9.0.0",
     "@nestjs/testing": "^9.0.0",
-    "@nrwl/angular": "^14.5.10",
-    "@nrwl/cli": "14.5.6",
-    "@nrwl/cypress": "14.5.6",
-    "@nrwl/eslint-plugin-nx": "14.5.6",
-    "@nrwl/jest": "14.5.6",
-    "@nrwl/linter": "14.5.6",
-    "@nrwl/nest": "14.5.6",
-    "@nrwl/node": "14.5.6",
-    "@nrwl/nx-cloud": "latest",
-    "@nrwl/storybook": "^14.5.6",
-    "@nrwl/workspace": "14.5.6",
+    "@nrwl/angular": "14.5.10",
+    "@nrwl/cli": "14.5.10",
+    "@nrwl/cypress": "14.5.10",
+    "@nrwl/eslint-plugin-nx": "14.5.10",
+    "@nrwl/jest": "14.5.10",
+    "@nrwl/linter": "14.5.10",
+    "@nrwl/nest": "14.5.10",
+    "@nrwl/node": "14.5.10",
+    "@nrwl/nx-cloud": "14.5.4",
+    "@nrwl/storybook": "14.5.10",
+    "@nrwl/workspace": "14.5.10",
     "@storybook/addon-essentials": "~6.5.9",
     "@storybook/angular": "~6.5.9",
     "@storybook/builder-webpack5": "~6.5.9",
@@ -79,7 +80,7 @@
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.0.0",
     "ng-packagr": "~14.1.0",
-    "nx": "14.5.6",
+    "nx": "14.5.10",
     "postcss": "^8.4.5",
     "postcss-import": "~14.1.0",
     "postcss-preset-env": "~7.5.0",
@@ -94,3 +95,4 @@
   "readme": "ERROR: No README data found!",
   "_id": "setbrain-dashboard@0.0.0"
 }
+
